+++ docs/import/manage_disaster_startup.js
@@ -27,55 +27,4 @@ $(() => {
   $('#cancel-new-disaster').on('click', () => toggleState(true));
   setUpScoreSelectorTable();
   taskAccumulator.taskCompleted();
-  initializeMap();
 });
-
-function initializeMap() {
-  const {map} = createBasicMap(document.getElementById('map-bounds-map'), {streetViewControl: false});
-  const drawingManager = new google.maps.drawing.DrawingManager({
-    drawingControl: true,
-    drawingControlOptions: {drawingModes: ['polygon']},
-    polygonOptions: {editable: true, draggable: true},
-  });
-
-  const deleteButton = $(document.createElement('button'));
-  let polygon;
-  drawingManager.addListener('overlaycomplete', (event) => {
-    polygon = event.overlay;
-    callbackOnPolygonChange(polygon);
-    drawingManager.setMap(null);
-    deleteButton.show();
-  });
-
-  drawingManager.setMap(map);
-  deleteButton.text('Delete');
-  deleteButton.hide();
-  deleteButton.on('click', () => {
-    polygon.setMap(null);
-    drawingManager.setMap(map);
-    drawingManager.setDrawingMode(null);
-    deleteButton.hide();
-  });
-  map.controls[google.maps.ControlPosition.TOP_LEFT].insertAt(0, deleteButton[0]);
-  return {map, drawingManager};
-}
-
-function callbackOnPolygonChange(polygon) {
-  const callback = () => console.log('here');
-  const listeners = addListenersToPolygon(polygon, callback);
-  polygon.addListener('dragend', () => {
-    listeners.push(...addListenersToPolygon(polygon, callback));
-    callback();
-  });
-  polygon.addListener('dragstart', () => {
-    listeners.forEach((listener) => google.maps.event.removeListener(listener));
-    listeners.length = 0;
-  });
-}
-
-const polygonPathEventTypes = ['insert_at', 'remove_at', 'set_at'];
-
-function addListenersToPolygon(polygon, callback) {
-  const path = polygon.getPath();
-  return polygonPathEventTypes.map((eventType) => google.maps.event.addListener(path, eventType, callback));
-}
